+ stats manager
we need to break apart the telegram bot manager class and this is a good place to start. I really not a fan of the current design of this solution but we'll take small steps to make it better
