+++ TheCountBot.Application/IStatsManager.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Threading.Tasks;
+using TheCountBot.Application.Models;
+
+namespace TheCountBot
+{
+    internal interface IStatsManager
+    {
+        Task HandleStatsCommandAsync( BotCommand command, string user, IServiceProvider serviceProvider );
+    }
+}
