+++ TheCountBot.Application/TelegramBotManager.cs
@@ -34,7 +34,9 @@ namespace TheCountBot
 
         private readonly INumberStoreRepository _numberStoreRepository;
 
-        public TelegramBotManager( IOptions<Settings> settingsOptions, ITelegramBotClient telegramBotClient, INumberStoreRepository numberStoreRepository )
+        private readonly IStatsManager _statsManager;
+
+        public TelegramBotManager( IOptions<Settings> settingsOptions, ITelegramBotClient telegramBotClient, INumberStoreRepository numberStoreRepository, IStatsManager statsManager )
         {
             _settings = settingsOptions.Value;
             _botClient = telegramBotClient;
@@ -42,6 +44,7 @@ namespace TheCountBot
             _stateTimer = new Timer( TimerFunc, null, _settings.TimerWaitTime, _settings.TimerWaitTime );
             _insultList = _settings.InsultsForMessingUpTheNumber;
             _numberStoreRepository = numberStoreRepository;
+            _statsManager = statsManager;
         }
 
         public async Task RunAsync( IServiceProvider serviceProvider )
@@ -66,109 +69,6 @@ namespace TheCountBot
             await _botClient.SendTextMessageAsync( _settings.MetaCountingChatId, message, mode );
         }
 
-        private async Task CalculateAndSendLimitedStatsPerPersonAsync( List<MessageEntry> list )
-        {
-            var totalMistakesByUser = getMistakesByUser(list);
-            var totalMessagesByUser = getMessagesByUser(list);
-
-            string messageToSend = String.Format( $"```\n{"Username", -20} -- {"Total Messages Sent", -20} -- {"Number Of Mistakes", -20} -- {"Error Rate", -20}\n" );
-            totalMistakesByUser.Keys.ToList().ForEach( username => {
-                int totalMessagesSent = totalMessagesByUser[username];
-                int totalMistakes = totalMistakesByUser[username];
-                double errorRate = ((double) totalMistakes) / totalMessagesSent * 100;
-
-                messageToSend += String.Format( $"{username, -20} -- {totalMessagesSent, -20} -- {totalMistakes, -20} -- {errorRate,-20:0.##}\n" );
-            } );
-            messageToSend += "```";
-
-            await SendMessageAsync( messageToSend, ParseMode.Markdown );
-        }
-
-        private async Task CalculateAndSendFullStatsPerPersonAsync( List<MessageEntry> list )
-        {
-            var totalMistakesByUser = getMistakesByUser(list);
-            var totalMessagesByUser = getMessagesByUser(list);
-            var mistakeRatesByUser = new Dictionary<String, double>();
-
-            int countOfTotalMistakes = totalMistakesByUser.Keys.ToList().Aggregate(0, (acc, key) => acc + totalMistakesByUser[key]);
-            int countOfTotalMessages = totalMessagesByUser.Keys.ToList().Aggregate(0, (acc, key) => acc + totalMessagesByUser[key]);
-
-            string messageToSend = String.Format( $"```\n{"Username",-20} -- {"Total Messages Sent",-20} -- {"Number Of Mistakes",-20} -- {"Error Rate",-20}\n" );
-            totalMistakesByUser.Keys.ToList().ForEach(username => {
-                int totalMessagesSent = totalMessagesByUser[username];
-                int totalMistakes = totalMistakesByUser[username];
-                double errorRate = ((double)totalMistakes) / totalMessagesSent * 100;
-                mistakeRatesByUser.Add(username, errorRate);
-
-                messageToSend += String.Format( $"{username,-20} -- {totalMessagesSent,-20} -- {totalMistakes,-20} -- {errorRate,-20:0.##}\n" );
-            });
-            messageToSend += "\n";
-            var relativeErrorRates = calculateRelativeMistakeRatesByUser(mistakeRatesByUser);
-            messageToSend += String.Format( $"\n{"Username",-20} -- {"Total Message Percentage",-30} -- {"Total Error Rate",-20} -- {"Relative Error Rate",-20}\n" );
-            totalMistakesByUser.Keys.ToList().ForEach(username => {
-                int totalMessagesSent = totalMessagesByUser[username];
-                int totalMistakes = totalMistakesByUser[username];
-                double totalMessagePercentage = ((double)totalMessagesSent) / countOfTotalMessages * 100;
-                double totalErrorRate = ((double)totalMistakes) / countOfTotalMistakes * 100;
-                double relativeError = relativeErrorRates[username];
-
-                messageToSend += String.Format( $"{username,-20} -- {totalMessagePercentage,-30:0.##} -- {totalErrorRate,-20:0.##} -- {relativeError,-20:0.##}\n" );
-            });
-            messageToSend += "```";
-
-            await SendMessageAsync(messageToSend, ParseMode.Markdown);
-        }
-
-        private async Task CalculateAndSendIndividualStats( List<MessageEntry> list, String username )
-        {
-            var totalMistakesByUser = getMistakesByUser( list );
-            var totalMessagesByUser = getMessagesByUser( list );
-
-            String messageToSend = "";
-            if( !totalMessagesByUser.ContainsKey( username ) )
-            {
-                messageToSend = "You haven't counted yet!";
-            }
-            else
-            {
-                messageToSend = BuildIndividualStatsMessage( totalMistakesByUser, totalMessagesByUser, username );
-            }
-
-            await SendMessageAsync( messageToSend, ParseMode.Markdown );
-        }
-
-        private String BuildIndividualStatsMessage( Dictionary<String, int> totalMistakesByUser, Dictionary<String, int> totalMessagesByUser, String username )
-        {
-            int countOfTotalMistakes = totalMistakesByUser.Keys.ToList().Aggregate(0, ( acc, key ) => acc + totalMistakesByUser[key]);
-
-            string messageToSend = String.Format( $"```\n{"Username",-20} -- {"Total Messages",-20} -- {"Number Of Mistakes",-20} -- {"Error Rate",-20} -- {"Total Error Percentage",-20}\n" );
-
-            int totalMessagesSent = totalMessagesByUser[username];
-            int totalMistakes = totalMistakesByUser[username];
-            double errorRate = ((double)totalMistakes) / totalMessagesSent * 100;
-            double totalErrorShare = ((double)totalMistakes) / countOfTotalMistakes * 100;
-
-            messageToSend += String.Format( $"{username,-20} -- {totalMessagesSent,-20} -- {totalMistakes,-20} -- {errorRate,-20:0.##} -- {totalErrorShare,-20:0.##}\n" );
-            messageToSend += "```";
-            return messageToSend;
-        }
-
-        private async Task HandleStatsCommandAsync( BotCommand command, String user )
-        {
-            switch( command.commandType )
-            {
-                case BotCommandEnum.fullStats:
-                    await CalculateAndSendFullStatsPerPersonAsync( await _numberStoreRepository.GetHistoryAsync( _serviceProvider ) );
-                    break;
-                case BotCommandEnum.limitedStats:
-                    await CalculateAndSendLimitedStatsPerPersonAsync( await _numberStoreRepository.GetHistoryAsync( _serviceProvider ) );
-                    break;
-                case BotCommandEnum.individualStats:
-                    await CalculateAndSendIndividualStats( await _numberStoreRepository.GetHistoryAsync( _serviceProvider ), user );
-                    break;
-            }
-        }
-
         private bool MoreRobustNumberCheck( string x )
         {
             if ( x.StartsWith( "0" ) ) return false;
@@ -186,7 +86,7 @@ namespace TheCountBot
                 BotCommand command = new BotCommand(e.Message.Text);
                 if (command.commandType != BotCommandEnum.noCommand)
                 {
-                    await HandleStatsCommandAsync( command, e.Message.From.Username );
+                    await _statsManager.HandleStatsCommandAsync( command, e.Message.From.Username, _serviceProvider );
                     return;
                 }
             }
@@ -311,64 +211,6 @@ namespace TheCountBot
                 await SendMessageAsync( $"@{user} blaze it!" );
             }
         }
-
-        private Dictionary<String,int> getMistakesByUser( List<MessageEntry> list )
-        {
-            var totalMistakesByUser = new Dictionary<String, int>();
-
-            foreach ( MessageEntry record in list )
-            {
-                if ( !totalMistakesByUser.ContainsKey( record.Username ) )
-                {
-                    totalMistakesByUser[record.Username] = 0;
-                }
-                if ( !record.Correct )
-                {
-                    totalMistakesByUser[record.Username] += 1;
-                }
-            }
-            return totalMistakesByUser;
-        }
-
-        private Dictionary<String,int> getMessagesByUser( List<MessageEntry> list )
-        {
-            var totalMessagesByUser = new Dictionary<String, int>();
-
-            foreach  ( MessageEntry record in list )
-            {
-                if ( !totalMessagesByUser.ContainsKey( record.Username ) )
-                {
-                    totalMessagesByUser[record.Username] = 0;
-                }
-                totalMessagesByUser[record.Username] += 1;
-            }
-            return totalMessagesByUser;
-        }
-
-        private Dictionary<String, double> calculateRelativeMistakeRatesByUser( Dictionary<String, double> mistakeRates )
-        {
-            Dictionary<String, double> relativeRates = new Dictionary<string, double>();
-            double minimumError = getPositiveMinimumNonZeroErrorRate( mistakeRates );
-            mistakeRates.Keys.ToList().ForEach( username =>
-            {
-                relativeRates.Add( username, mistakeRates[username] / minimumError );
-            });
-            return relativeRates;
-        }
-
-        private double getPositiveMinimumNonZeroErrorRate( Dictionary<String, double> mistakeRates )
-        {
-            double minimumErrorPercentage = 101;
-            mistakeRates.Keys.ToList().ForEach (username =>
-            {
-                if( mistakeRates[username] < minimumErrorPercentage && mistakeRates[username] > 0.0001 )
-                {
-                    minimumErrorPercentage = mistakeRates[username];
-                }
-            }) ;
-            return minimumErrorPercentage;
-        }
-
         private string GetRandomInsultMessageForUser( string user )
         {
             int _randInt = _rng.Next( 0, _insultList.Count );
